+++ src/game/propobj.c
@@ -8870,7 +8870,7 @@ bool propobjFindHit(Model *model, ModelNode *startNode, coord3d *rayPos, coord3d
                     {
                         if (rwdata->gdl == rodata->Primary)
                         {
-                            s3 = (Gfx *)((uintptr_t)rodata->BaseAddr + ((u32)rodata->Primary & 0xffffff));
+                            s3 = (Gfx *)((uintptr_t)rodata->BaseAddr + ((u32)(uintptr_t)rodata->Primary & 0xffffff));
                         }
                         else
                         {
@@ -8879,7 +8879,7 @@ bool propobjFindHit(Model *model, ModelNode *startNode, coord3d *rayPos, coord3d
 
                         if (rodata->Secondary != NULL)
                         {
-                            s5 = (void *)((uintptr_t)rodata->BaseAddr + ((u32)rodata->Secondary & 0xffffff));
+                            s5 = (void *)((uintptr_t)rodata->BaseAddr + ((u32)(uintptr_t)rodata->Secondary & 0xffffff));
                         }
 
                         vertices = rwdata->Vertices;
@@ -8893,11 +8893,11 @@ bool propobjFindHit(Model *model, ModelNode *startNode, coord3d *rayPos, coord3d
 
                     if (rodata->Primary != NULL)
                     {
-                        s3 = (Gfx *)((uintptr_t)rodata->BaseAddr + ((u32)rodata->Primary & 0xffffff));
+                        s3 = (Gfx *)((uintptr_t)rodata->BaseAddr + ((u32)(uintptr_t)rodata->Primary & 0xffffff));
 
                         if (rodata->Secondary != NULL)
                         {
-                            s5 = (Gfx *)((uintptr_t)rodata->BaseAddr + ((u32)rodata->Secondary & 0xffffff));
+                            s5 = (Gfx *)((uintptr_t)rodata->BaseAddr + ((u32)(uintptr_t)rodata->Secondary & 0xffffff));
                         }
 
                         vertices = (void *)(uintptr_t)rodata->BaseAddr;
